+++ src/sdk/conference/channel.js
@@ -34,6 +34,7 @@ export class ConferencePeerConnectionChannel extends EventDispatcher {
     super();
     this._config = config;
     this._options = null;
+    this._videoCodecs = undefined;
     this._signaling = signaling;
     this._pc = null;
     this._internalId = null; // It's publication ID or subscription ID.
@@ -75,13 +76,20 @@ export class ConferencePeerConnectionChannel extends EventDispatcher {
     }
   }
 
-  publish(stream, options) {
+  publish(stream, options, videoCodecs) {
     if (options === undefined) {
       options = {audio: !!stream.mediaStream.getAudioTracks().length, video: !!stream
           .mediaStream.getVideoTracks().length};
     }
     if (typeof options !== 'object') {
       return Promise.reject(new TypeError('Options should be an object.'));
+    }
+    if ((this._isRtpEncodingParameters(options.audio) &&
+         this._isOwtEncodingParameters(options.video)) ||
+        (this._isOwtEncodingParameters(options.audio) &&
+         this._isRtpEncodingParameters(options.video))) {
+      return Promise.reject(new ConferenceError(
+          'Mixing RTCRtpEncodingParameters and AudioEncodingParameters/VideoEncodingParameters is not allowed.'))
     }
     if (options.audio === undefined) {
       options.audio = !!stream.mediaStream.getAudioTracks().length;
@@ -115,18 +123,20 @@ export class ConferencePeerConnectionChannel extends EventDispatcher {
         }
       }
     }
-    if (typeof options.video === 'object') {
-      if (!Array.isArray(options.video)) {
-        return Promise.reject(new TypeError(
-            'options.video should be a boolean or an array.'));
-      }
+    if (typeof options.video === 'object' && !Array.isArray(options.video)) {
+      return Promise.reject(new TypeError(
+        'options.video should be a boolean or an array.'));
+    }
+    if (this._isOwtEncodingParameters(options.video)) {
       for (const parameters of options.video) {
-        if (!parameters.codec || typeof parameters.codec.name !== 'string' || (
-          parameters.maxBitrate !== undefined && typeof parameters.maxBitrate
-          !== 'number') || (parameters.codec.profile !== undefined
-          && typeof parameters.codec.profile !== 'string')) {
+        if (!parameters.codec || typeof parameters.codec.name !== 'string' ||
+          (
+            parameters.maxBitrate !== undefined && typeof parameters
+            .maxBitrate !==
+            'number') || (parameters.codec.profile !== undefined &&
+            typeof parameters.codec.profile !== 'string')) {
           return Promise.reject(new TypeError(
-              'options.video has incorrect parameters.'));
+            'options.video has incorrect parameters.'));
         }
       }
     }
@@ -149,9 +159,6 @@ export class ConferencePeerConnectionChannel extends EventDispatcher {
       }
       mediaOptions.audio = {};
       mediaOptions.audio.source = stream.source.audio;
-      for (const track of stream.mediaStream.getAudioTracks()) {
-        this._pc.addTrack(track, stream.mediaStream);
-      }
     } else {
       mediaOptions.audio = false;
     }
@@ -174,9 +181,6 @@ export class ConferencePeerConnectionChannel extends EventDispatcher {
         },
         framerate: trackSettings.frameRate,
       };
-      for (const track of stream.mediaStream.getVideoTracks()) {
-        this._pc.addTrack(track, stream.mediaStream);
-      }
     } else {
       mediaOptions.video = false;
     }
@@ -193,11 +197,28 @@ export class ConferencePeerConnectionChannel extends EventDispatcher {
       this._internalId = data.id;
       if (typeof this._pc.addTransceiver === 'function') {
         // |direction| seems not working on Safari.
-        if (mediaOptions.audio && stream.mediaStream.getAudioTracks() > 0) {
-          this._pc.addTransceiver('audio', {direction: 'sendonly'});
+        if (mediaOptions.audio && stream.mediaStream.getAudioTracks().length >
+          0) {
+          const transceiverInit = {
+            direction: 'sendonly'
+          };
+          if (this._isRtpEncodingParameters(options.audio)) {
+            transceiverInit.sendEncodings = options.audio;
+          }
+          this._pc.addTransceiver(stream.mediaStream.getAudioTracks()[0],
+            transceiverInit);
         }
-        if (mediaOptions.video && stream.mediaStream.getVideoTracks() > 0) {
-          this._pc.addTransceiver('video', {direction: 'sendonly'});
+        if (mediaOptions.video && stream.mediaStream.getVideoTracks().length >
+          0) {
+          const transceiverInit = {
+            direction: 'sendonly'
+          };
+          if (this._isRtpEncodingParameters(options.video)) {
+            transceiverInit.sendEncodings = options.video;
+            this._videoCodecs = videoCodecs;
+          }
+          this._pc.addTransceiver(stream.mediaStream.getVideoTracks()[0],
+            transceiverInit);
         }
       }
       let localDesc;
@@ -235,18 +256,18 @@ export class ConferencePeerConnectionChannel extends EventDispatcher {
   subscribe(stream, options) {
     if (options === undefined) {
       options = {
-        audio: !!stream.capabilities.audio,
-        video: !!stream.capabilities.video,
+        audio: !!stream.settings.audio,
+        video: !!stream.settings.video,
       };
     }
     if (typeof options !== 'object') {
       return Promise.reject(new TypeError('Options should be an object.'));
     }
     if (options.audio === undefined) {
-      options.audio = !!stream.capabilities.audio;
+      options.audio = !!stream.settings.audio;
     }
     if (options.video === undefined) {
-      options.video = !!stream.capabilities.video;
+      options.video = !!stream.settings.video;
     }
     if ((options.audio !== undefined && typeof options.audio !== 'object' &&
         typeof options.audio !== 'boolean' && options.audio !== null) || (
@@ -254,8 +275,8 @@ export class ConferencePeerConnectionChannel extends EventDispatcher {
         typeof options.video !== 'boolean' && options.video !== null)) {
       return Promise.reject(new TypeError('Invalid options type.'));
     }
-    if (options.audio && !stream.capabilities.audio || (options.video &&
-        !stream.capabilities.video)) {
+    if (options.audio && !stream.settings.audio || (options.video &&
+        !stream.settings.video)) {
       return Promise.reject(new ConferenceError(
           'options.audio/video cannot be true or an object if there is no '
           + 'audio/video track in remote stream.'
@@ -299,6 +320,7 @@ export class ConferencePeerConnectionChannel extends EventDispatcher {
           bitrate: options.video.bitrateMultiplier ? 'x'
               + options.video.bitrateMultiplier.toString() : undefined,
           keyFrameInterval: options.video.keyFrameInterval,
+          simulcastRid: options.video.rid
         };
       }
     } else {
@@ -637,11 +659,25 @@ export class ConferencePeerConnectionChannel extends EventDispatcher {
   }
 
   _setRtpSenderOptions(sdp, options) {
+    // SDP mugling is deprecated, moving to `setParameters`.
+    if (this._isRtpEncodingParameters(options.audio) ||
+        this._isRtpEncodingParameters(options.video)) {
+      return sdp;
+    }
     sdp = this._setMaxBitrate(sdp, options);
     return sdp;
   }
 
   _setRtpReceiverOptions(sdp, options) {
+    // _videoCodecs is a workaround for setting video codecs. It will be moved to RTCRtpSendParameters.
+    if (this._isRtpEncodingParameters(options.video) && this._videoCodecs) {
+      sdp = SdpUtils.reorderCodecs(sdp, 'video', this._videoCodecs);
+      return sdp;
+    }
+    if (this._isRtpEncodingParameters(options.audio) ||
+        this._isRtpEncodingParameters(options.video)) {
+      return sdp;
+    }
     sdp = this._setCodecOrder(sdp, options);
     return sdp;
   }
@@ -674,5 +710,24 @@ export class ConferencePeerConnectionChannel extends EventDispatcher {
     } else {
       Logger.warning('Invalid data value for stream update info.');
     }
+  }
+
+  _isRtpEncodingParameters(obj) {
+    if (!Array.isArray(obj)) {
+      return false;
+    }
+    // Only check the first one.
+    const param = obj[0];
+    return param.codecPayloadType || param.dtx || param.active || param
+      .ptime || param.maxFramerate || param.scaleResolutionDownBy || param.rid;
+  }
+
+  _isOwtEncodingParameters(obj) {
+    if (!Array.isArray(obj)) {
+      return false;
+    }
+    // Only check the first one.
+    const param = obj[0];
+    return !!param.codec;
   }
 }