+++ docs/mdfiles/changelog.md
@@ -1,5 +1,10 @@
 Change Log
 ==========
+# 4.3
+* The type of `conferenceClient.publish` method's option is changed from `AudioEncodingParameters` and `VideoEncodingParameters` to `RTCRtpSendParameters`.
+* `audio` and `video` of `PublicationSettings` is changed from `AudioPublicationSettings` and `VideoPublicationSettings` to `AudioPublicationSettings` array and `VideoPublicationSettings` array.
+* Add `rid` to `VideoSubscriptionConstraints`. When `rid` is specified, other constraints will be ignored.
+
 # 4.0.2
 * Fix a compatibility issue for Chrome 69 beta.